ComponentApiDoc PERF refactoring WIP
