+++ Havit.Blazor.Components.Web.Bootstrap.Documentation/Havit.Blazor.Components.Web.Bootstrap.Documentation.csproj
@@ -36,19 +36,19 @@
 
 	<ItemGroup>
 		<Content Remove="**/*.CodeSnippet.*" />
+		<Content Remove="GenericTypePlaceholders\**" />
 		<Compile Remove="**/*.CodeSnippet.*" />
+		<Compile Remove="GenericTypePlaceholders\**" />
 		<EmbeddedResource Include="**/*.CodeSnippet.*" />
+		<EmbeddedResource Remove="GenericTypePlaceholders\**" />
+		<None Remove="GenericTypePlaceholders\**" />
 		<EmbeddedResource Update="**/*.CodeSnippet.*">
 			<CopyToOutputDirectory>Never</CopyToOutputDirectory>
 		</EmbeddedResource>
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Folder Include="GenericTypePlaceholders\" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <None Include="Pages\Components\HxMultiSelectDoc\HxMultiSelect_Documentation.razor" />
+		<None Include="Pages\Components\HxMultiSelectDoc\HxMultiSelect_Documentation.razor" />
 	</ItemGroup>
 
 	<Target Name="EmbedDemos" BeforeTargets="PrepareForBuild">
