+++ Havit.Blazor.Components.Web.Bootstrap.Documentation/Services/IComponentApiDocModelBuilder.cs
@@ -1,0 +1,7 @@
+﻿using Havit.Blazor.Components.Web.Bootstrap.Documentation.Model;
+
+namespace Havit.Blazor.Components.Web.Bootstrap.Documentation.Services;
+public interface IComponentApiDocModelBuilder
+{
+	ComponentApiDocModel BuildModel(Type type, bool isDelegate, List<string> excludedMembers);
+}
