+++ Havit.Blazor.Components.Web.Bootstrap.Documentation/Shared/Components/ComponentApiDoc.razor
@@ -1,23 +1,24 @@
-﻿<h1 id="@RemoveSpecialCharacters(Type.Name)">@RemoveSpecialCharacters(Type.Name)</h1>
-@if (!String.IsNullOrWhiteSpace(classMember?.Comments?.Summary))
+﻿@using Havit.Blazor.Components.Web.Bootstrap.Documentation.Services
+<h1 id="@ApiRenderer.RemoveSpecialCharacters(Type.Name)">@ApiRenderer.RemoveSpecialCharacters(Type.Name)</h1>
+@if (!String.IsNullOrWhiteSpace(model.Class?.Comments?.Summary))
 {
-	<p class="lead">@((MarkupString)classMember.Comments.Summary)</p>
+	<p class="lead">@((MarkupString)model.Class.Comments.Summary)</p>
 }
 
 @ChildContent
 @MainContent
 
-@if (parameters is not null || properties is not null || events is not null || staticMethods is not null || staticProperties is not null || CssVariables is not null)
+@if (model.Parameters is not null || model.Properties is not null || model.Events is not null || model.StaticMethods is not null || model.StaticProperties is not null || CssVariables is not null)
 {
 	<h2>API</h2>
 }
 
 @if (Delegate)
 {
-	<h6><code>@((MarkupString)delegateSignature)</code></h6>
+	<h6><code>@((MarkupString)model.DelegateSignature)</code></h6>
 }
 
-@if (isEnum)
+@if (model.IsEnum)
 {
 	<h3>Enum Values</h3>
 	<div class="table-responsive">
@@ -30,7 +31,7 @@
 				</tr>
 			</thead>
 			<tbody>
-				@foreach (var enumMember in enumMembers)
+				@foreach (var enumMember in model.EnumMembers)
 				{
 					<tr>
 						<td>@enumMember.Index</td>
@@ -43,7 +44,7 @@
 	</div>
 }
 
-@if (parameters.Count > 0)
+@if (model.Parameters.Any())
 {
 	<h3>Parameters</h3>
 
@@ -57,7 +58,7 @@
 				</tr>
 			</thead>
 			<tbody>
-				@foreach (var property in parameters.OrderBy(p => p.PropertyInfo.Name))
+				@foreach (var property in model.Parameters.OrderBy(p => p.PropertyInfo.Name))
 				{
 					<tr>
 						<td>
@@ -67,7 +68,7 @@
 									@("REQUIRED")
 								</HxBadge>
 							}
-							@if (IsPropertyStatic(property))
+							@if (property.IsStatic)
 							{
 								<HxBadge Color="ThemeColor.Light" TextColor="ThemeColor.Dark">
 									@("static")
@@ -76,7 +77,7 @@
 
 							<strong>@property.PropertyInfo.Name</strong>
 						</td>
-						<td><code>@((MarkupString)@FormatType(property.PropertyInfo.PropertyType))</code></td>
+						<td><code>@((MarkupString)ApiRenderer.FormatType(property.PropertyInfo.PropertyType))</code></td>
 						<td class="summary">@((MarkupString)property.Comments.Summary)</td>
 					</tr>
 				}
@@ -85,7 +86,7 @@
 	</div>
 }
 
-@if (properties.Count > 0)
+@if (model.Properties.Any())
 {
 	<h3>Other properties</h3>
 	<div class="table-responsive">
@@ -98,11 +99,11 @@
 				</tr>
 			</thead>
 			<tbody>
-				@foreach (var property in properties.OrderBy(p => p.PropertyInfo.Name))
+				@foreach (var property in model.Properties.OrderBy(p => p.PropertyInfo.Name))
 				{
 					<tr>
 						<td><strong>@property.PropertyInfo.Name</strong></td>
-						<td><code>@((MarkupString)@FormatType(property.PropertyInfo.PropertyType))</code></td>
+						<td><code>@((MarkupString)ApiRenderer.FormatType(property.PropertyInfo.PropertyType))</code></td>
 						<td class="summary">@((MarkupString)property.Comments.Summary)</td>
 					</tr>
 				}
@@ -111,7 +112,7 @@
 	</div>
 }
 
-@if (events.Count > 0)
+@if (model.Events.Any())
 {
 	<h3>Event callbacks</h3>
 
@@ -125,11 +126,11 @@
 				</tr>
 			</thead>
 			<tbody>
-				@foreach (var currentEvent in events.OrderBy(e => e.PropertyInfo.Name))
+				@foreach (var currentEvent in model.Events.OrderBy(e => e.PropertyInfo.Name))
 				{
 					<tr>
 						<td><strong>@currentEvent.PropertyInfo.Name</strong> </td>
-						<td><code>@((MarkupString)@FormatType(currentEvent.PropertyInfo.PropertyType))</code></td>
+						<td><code>@((MarkupString)ApiRenderer.FormatType(currentEvent.PropertyInfo.PropertyType))</code></td>
 						<td class="summary">@((MarkupString)currentEvent.Comments.Summary)</td>
 					</tr>
 				}
@@ -138,7 +139,7 @@
 	</div>
 }
 
-@if (methods.Count > 0)
+@if (model.Methods.Any())
 {
 	<h3>Methods</h3>
 	<div class="table-responsive">
@@ -151,11 +152,11 @@
 				</tr>
 			</thead>
 			<tbody>
-				@foreach (var method in methods.OrderBy(m => m.MethodInfo.Name))
+				@foreach (var method in model.Methods.OrderBy(m => m.MethodInfo.Name))
 				{
 					<tr>
 						<td><strong>@method.MethodInfo.Name</strong>@((MarkupString)@method.GetParameters())</td>
-						<td><code>@((MarkupString)@FormatMethod(method.MethodInfo.ReturnType))</code></td>
+						<td><code>@((MarkupString)ApiRenderer.FormatMethod(method.MethodInfo.ReturnType, model))</code></td>
 						<td class="summary">@((MarkupString)method.Comments.Summary)</td>
 					</tr>
 				}
@@ -164,7 +165,7 @@
 	</div>
 }
 
-@if (staticProperties.Count > 0)
+@if (model.StaticProperties.Any())
 {
 	<h3>Static properties</h3>
 
@@ -178,11 +179,11 @@
 				</tr>
 			</thead>
 			<tbody>
-				@foreach (var property in staticProperties.OrderBy(p => p.PropertyInfo.Name))
+				@foreach (var property in model.StaticProperties.OrderBy(p => p.PropertyInfo.Name))
 				{
 					<tr>
 						<td><strong>@property.PropertyInfo.Name</strong></td>
-						<td><code>@((MarkupString)@FormatType(property.PropertyInfo.PropertyType))</code></td>
+						<td><code>@((MarkupString)ApiRenderer.FormatType(property.PropertyInfo.PropertyType))</code></td>
 						<td class="summary">@((MarkupString)property.Comments.Summary)</td>
 					</tr>
 				}
@@ -191,7 +192,7 @@
 	</div>
 }
 
-@if (staticMethods.Count > 0)
+@if (model.StaticMethods.Any())
 {
 	<h3>Static Methods</h3>
 
@@ -205,11 +206,11 @@
 				</tr>
 			</thead>
 			<tbody>
-				@foreach (var method in staticMethods.OrderBy(m => m.MethodInfo.Name))
+				@foreach (var method in model.StaticMethods.OrderBy(m => m.MethodInfo.Name))
 				{
 					<tr>
 						<td><strong>@method.MethodInfo.Name</strong> @((MarkupString)@method.GetParameters())</td>
-						<td><code>@((MarkupString)@FormatType(method.MethodInfo.ReturnType))</code></td>
+						<td><code>@((MarkupString)ApiRenderer.FormatType(method.MethodInfo.ReturnType))</code></td>
 						<td class="summary">@((MarkupString)method.Comments.Summary)</td>
 					</tr>
 				}
